revert back to x11 from wayland
We are getting reports of linux users unable to use podman desktop now,
so we are reverting back to x11 which has been consistent in not having
startup failures.

As per flatpak docs we shouldn't use wayland just yet: "Native wayland
support in electron is experimental and often unstable. It is advised to
stick with the X11 and Xwayland configuration above as the default."
(https://docs.flatpak.org/en/latest/electron.html)

Reverting back to x11.

Closes https://github.com/flathub/io.podman_desktop.PodmanDesktop/issues/51

Signed-off-by: Charlie Drage <charlie@charliedrage.com>